Bug-Fix: removed possible nullpointer exception
git-svn-id: https://svn.java.net/svn/docsources~svn/trunk@3419 3b25252b-017b-4ae5-8f61-fa87a855a72e

+++ DocSources/src/main/java/org/medici/bia/service/teaching/TeachingServiceImpl.java
@@ -1506,7 +1506,7 @@ public class TeachingServiceImpl implements TeachingService {
 		
 		// if mails (for subscribed users) are still not sent, the forum post notification is removed
 		ForumPostNotified forumPostNotified = getForumPostNotifiedDAO().getForumPostNotifiedByPost(post.getPostId());
-		if (Boolean.FALSE.equals(forumPostNotified.getMailSended())) {
+		if (forumPostNotified != null && Boolean.FALSE.equals(forumPostNotified.getMailSended())) {
 			getForumPostNotifiedDAO().remove(forumPostNotified);
 		}
